backend/requirements.txt
Flask==2.3.3
pandas==2.1.1
flask-cors==4.0.0
werkzeug==2.3.7
langchain==0.1.0
langchain-anthropic>=0.1.1
anthropic>=0.18.0
python-dotenv==1.0.0